+++ webapp/src/components/Fragments/Footer.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { makeStyles } from '@material-ui/core/styles';
-import { color } from "@mui/system";
 import { Typography } from "@mui/material";
+
 const useStyles = makeStyles({
     stickToBottom: {
       backgroundColor:'#000000',
@@ -13,20 +13,17 @@ const useStyles = makeStyles({
       left: '0',
       width: '100%',
       textAlign:'center'
-      
     },
-    
+});
 
-  });
 const Footer = () => {
     const classes = useStyles();
     return (
-  <div className={classes.stickToBottom}>
-    <Typography variant='h6' >
-        @DeDe_es3A
-    </Typography>
-  </div>
+      <div className={classes.stickToBottom}>
+        <Typography variant='h6' >
+          @DeDe_es3A
+        </Typography>
+      </div>
     );
 };
-
 export default Footer;
